requirements.txt
biopython>=1.83
pandas>=2.0
numpy>=1.24
torch>=2.0
transformers>=4.35
requests>=2.31
tqdm>=4.66
matplotlib>=3.7
seaborn>=0.13
scipy>=1.11
